Move new reference mapping to a separate component
