+++ src/integrate/scvi_tools/config.vsh.yaml
@@ -1,45 +1,62 @@
 functionality:
-  name: scarches
+  name: scvi
   namespace: "integrate"
   version: "dev"
-  description: "Performs reference mapping with scArches"
+  description: "Performs scvi integration as done in the human lung cell atlas https://github.com/LungCellAtlas/HLCA"
   authors:
-    - name: Vladimir Shitov
-      email: vladimir.shitov@helmholtz-muenchen.de
+    - name: Malte Luecken
+      email: malte.luecken@helmholtz-muenchen.de
       roles: [ author ]
-      props: { github: vladimirshitov, orcid: "0000-0002-1960-8812" }
+      props: { github: luckymd, orcid: "0000-0001-7464-7921" }
+    - name: Dries Schaumont
+      email: dries@data-intuitive.com
+      roles: [ maintainer ]
+      props: { github: DriesSchaumont, orcid: "0000-0002-4389-0440" }
   argument_groups:
     - name: Inputs
       arguments:
         - name: "--input"
           alternatives: ["-i"]
           type: file
-          description: Input h5mu file to use as a query
+          description: Input h5mu file
           direction: input
           required: true
-        - name: "--modality"
+        - name: "--base_model"
+          choices: ["scvi", "scanvi", "totalvi"]
+          alternatives: ["-m"]
+          type: string
+          description: Model to use for the integration
+          default: "scvi"
+          required: false
+        - name: "--query_modality"
           type: string
           default: "rna"
           required: false
-        - name: "--reference"
-          choices: ["HLCA"]
-          alternatives: ["-r"]
+        - name: "--query_proteins_key"
           type: string
-          description: Name of the reference dataset
+          description: Name of the modality containing proteins in the query
+          required: false
+        - name: "--reference_modality"
+          type: string
+          default: "rna"
+          required: false
+        - name: "--reference_proteins_key"
+          type: string
+          description: Name of the modality containing proteins in the reference
           required: false
         - name: "--unlabeled_category"
           type: string
-          description: "Value used for unlabeled cells"
+          description: "Value used for unlabeled cells. Used by scanvi base model"
           default: "Unknown"
           required: false
         - name: "--labels_key"
           type: string
-          description: "Name of the column containing cell types"
+          description: "Name of the column containing cell types. Used by scanvi base model"
           default: "leiden"
           required: false
         - name: "--predicted_labels_key"
           type: string
-          description: Name of the column that will contain cell types predicted by scANVI
+          description: Name of the column that will contain cell types predicted by scanvi base model
           default: "predicted_labels"
           required: false
         - name: "--input_layer"
@@ -65,7 +82,7 @@ functionality:
           required: true
         - name: "--obsm_output"
           type: string
-          default: "X_integrated_scanvi"
+          default: "X_integrated_{base_model}"
           required: false
           description: "In which .obsm slot to store the resulting integrated embedding."
     - name: "Early stopping arguments"
@@ -132,9 +149,7 @@ platforms:
           - anndata~=0.8.0
           - torchmetrics~=0.6.0
           - scvi-tools
-          - pandas
   - type: nextflow
     variant: vdsl3
     directives:
       label: [highmem, highcpu]
-  - type: native
